+++ myokit/formats/axon/_abf.py
@@ -254,7 +254,6 @@ class AbfFile(myokit.formats.SweepSource):
         # Read the protocol information
         self._n_adc = None
         self._n_dac = None
-
         self._rate = None
         self._mode = None
 
@@ -264,6 +263,7 @@ class AbfFile(myokit.formats.SweepSource):
         self._run_start_to_start = None
         self._sweeps_per_run = None
         self._sweep_start_to_start = None
+        self._equal_length_sweeps = None
         self._samples_per_channel = None
 
         # To be able to treat v1 and v2 slightly more easily, we define 3
@@ -281,7 +281,6 @@ class AbfFile(myokit.formats.SweepSource):
         # will have A/D but no (or no supported) D/A. Conversely protocol files
         # will have D/A only. So all in one sweep is easiest.
         self._sweeps = None
-
         self._read_3_protocol_information()
 
         # Read and calculate conversion factors for integer data in ADC
@@ -553,7 +552,10 @@ class AbfFile(myokit.formats.SweepSource):
             # 1.x versions only seem to have 1 DAC channel, but this is not
             # supported here.
             self._n_adc = int(h['nADCNumChannels'])
-            self._n_dac = min(len(h['sDACChannelName']), 2)
+            if h['sDACChannelName'] is None:    # pragma: no cover
+                self._n_dac = 2
+            else:
+                self._n_dac = min(len(h['sDACChannelName']), 2)
             self._rate = 1e6 / (h['fADCSampleInterval'] * self._n_adc)
             self._mode = h['nOperationMode']
         else:
@@ -577,11 +579,7 @@ class AbfFile(myokit.formats.SweepSource):
                 'Unsupported acquisition method'
                 f' {acquisition_modes[self._mode]}; unable to read D/A'
                 ' channels.')
-
-            # Remaining code is all about reading D/A info for episodic
-            # stimulation, so return
             self._n_dac = 0
-            return
 
         # Gather protocol information
         if self._version < 2:
@@ -668,9 +666,16 @@ class AbfFile(myokit.formats.SweepSource):
             self._sweep_start_to_start = self._samples_per_channel / self._rate
 
         # Create empty sweeps
-        n = h['lActualSweeps']
-        if self._is_protocol_file:
+        if self._mode == ACMODE_GAP_FREE:   # pragma: no cover
+            # In gap-free mode, treat the data as a single sweep
+            n = 1
+        elif self._is_protocol_file:
+            # In protocol file, there is no number of recorded sweeps, so use
+            # intended number of sweeps instead
             n = self._sweeps_per_run
+        else:
+            # Use number of recorded sweeps
+            n = h['lActualSweeps']
         self._sweeps = [Sweep() for i in range(n)]
 
         # User lists are not supported for D/A reconstruction
@@ -811,7 +816,8 @@ class AbfFile(myokit.formats.SweepSource):
         # Get binary integer format
         dt = np.dtype('i2') if h['nDataFormat'] == 0 else np.dtype('f4')
 
-        # Get number of channels, create a numpy memory map
+        # Get number of channels, create a numpy memory map of all data
+        # The map has size n and starts at offset o
         if self._version < 2:
             # Old files, get info from fields stored directly in header
             o = h['lDataSectionPtr'] * BLOCKSIZE \
@@ -823,37 +829,57 @@ class AbfFile(myokit.formats.SweepSource):
             n = h['sections']['Data']['length']
         data = np.memmap(self._filepath, dt, 'r', shape=(n,), offset=o)
 
-        # Load list of sweeps (Sweeps are called 'episodes' in ABF < 2)
+        # Load information about sweep sizes from "SynchArray".
+        # The number of sweep information structs will be n, and information
+        # starts at an offset o
         if self._version < 2:
             n = h['lSynchArraySize']
             o = h['lSynchArrayPtr'] * BLOCKSIZE
         else:
             n = h['sections']['SynchArray']['length']
             o = h['sections']['SynchArray']['index'] * BLOCKSIZE
-        if n > 0:
-            dt = [(str('offset'), str('i4')), (str('len'), str('i4'))]
-            sweep_data = np.memmap(
-                self._filepath, dt, 'r', shape=(n,), offset=o)
+
+        # Get sweep start times (real times, not array offsets) and sizes
+        # (as number of points in an array)
+        if self._mode != ACMODE_GAP_FREE:
+            # This will be read in as an array of numpy.Void objects, which are
+            # similar to structs. In ``dt`` we name the entries offset and len.
+            sweep_starts_sizes = np.memmap(
+                self._filepath, [('offset', 'i4'), ('len', 'i4')], 'r',
+                shape=(n,), offset=o)
         else:   # pragma: no cover
-            # Cover pragma: Don't have appropriate test file
-            sweep_data = np.empty((1), dt)
-            sweep_data[0]['len'] = data.size
-            sweep_data[0]['offset'] = 0
+            # In gap-free model, all data is a single sweep
+            sweep_starts_sizes = [(0, len(data))]
 
-        # Number of sweeps must equal n
-        if n != h['lActualSweeps']:
-            raise NotImplementedError(
-                'Unable to read file with different sizes per sweep.')
+        # Sanity check
+        if len(sweep_starts_sizes) != len(self._sweeps):  # pragma: no cover
+            raise RuntimeError(
+                'Unexpected situation: number of sweeps is'
+                f' {len(self._sweeps)} but found list of'
+                f' {len(sweep_starts_sizes)} sweep sizes.')
 
-        # Time-offset at start of first sweep
-        start = sweep_data[0]['offset'] / rate
+        # Check if sweep sizes are all the same
+        self._equal_length_sweeps = True
+        if len(sweep_starts_sizes) > 0:
+            size = sweep_starts_sizes[0][1]
+            for o, s in sweep_starts_sizes[1:]:
+                if s != size:   # pragma: no cover
+                    self._equal_length_sweeps = False
+                    break
+            del size
+
+        # Episodic stimulation? Then increment start time using sweep interval
+        # First time is time listed in first sweep
+        if self._mode == ACMODE_EPISODIC_STIMULATION:
+            tstart = sweep_starts_sizes[0][0] / rate
 
         # Get data
         pos = 0
-        for i_sweep, sdat in enumerate(sweep_data):
+        for i_sweep, (start, size) in enumerate(sweep_starts_sizes):
 
-            # Get the number of data points
-            size = sdat['len']
+            # Not episodic: Then use listed start time
+            if self._mode != ACMODE_EPISODIC_STIMULATION:   # pragma: no cover
+                tstart = start / rate
 
             # Calculate the correct size for variable-length event mode
             if self._mode == ACMODE_VARIABLE_LENGTH_EVENTS:  # pragma: no cover
@@ -874,25 +900,20 @@ class AbfFile(myokit.formats.SweepSource):
             # If needed, reformat the integers
             if h['nDataFormat'] == 0:
                 # Data given as integers? Convert to floating point
-
                 for i in range(self._n_adc):
                     part[:, i] *= self._adc_factors[i]
                     part[:, i] += self._adc_offsets[i]
 
-            # Get start in other modes
-            if self._mode != ACMODE_EPISODIC_STIMULATION:  # pragma: no cover
-                # All modes except episodic stimulation
-                start = data['offset'] / rate
-
-            # Create and populate sweep
+            # Create and populate sweep object
             sweep = self._sweeps[i_sweep]
             for i in range(self._n_adc):
                 c = Channel(self)
                 c._data = part[:, i]    # Actually store the data
                 c._rate = rate
-                c._start = start
+                c._start = tstart
                 c._is_reconstruction = False
 
+                # Add meta data
                 if self._version < 2:
                     j = h['nADCSamplingSeq'][i]
 
@@ -947,9 +968,9 @@ class AbfFile(myokit.formats.SweepSource):
 
                 sweep._channels.append(c)
 
+            # In episodic mode, increase time according to sweep interval
             if self._mode == ACMODE_EPISODIC_STIMULATION:
-                # Increase time according to sweeps in episodic stim. mode
-                start += self._sweep_start_to_start
+                tstart += self._sweep_start_to_start
 
     def _read_6_da_reconstructions(self):
         """
@@ -1033,6 +1054,12 @@ class AbfFile(myokit.formats.SweepSource):
                 raise IndexError(f'channel_id out of range: {channel_id}')
 
         return int_id
+
+    def acquisition_mode_str(self):
+        """
+        Returns the acquisition mode, as a string.
+        """
+        return acquisition_modes[self._mode]
 
     def channel(self, channel_id, join_sweeps=False):
         # Docstring in SweepSource
@@ -1193,8 +1220,7 @@ class AbfFile(myokit.formats.SweepSource):
         return self._da_units[index]
 
     def equal_length_sweeps(self):
-        # Always true for ABF
-        return True
+        return self._equal_length_sweeps
 
     def filename(self):
         """ Returns this ABF file's filename. """
@@ -1530,13 +1556,13 @@ HEADER_FIELDS = {
     # Note that a lot of the groups in the version 1 header start with obsolete
     # fields, followed later by their newer equivalents.
     1: [
-        ('fFileSignature', 0, '4s'),       # Coarse file version indication
+        ('fFileSignature', 0, '4s'),        # Coarse file version indication
         # Group 1, File info and sizes
-        ('fFileVersionNumber', 4, 'f'),    # Version number as float
-        ('nOperationMode', 8, 'h'),        # Acquisition mode
+        ('fFileVersionNumber', 4, 'f'),     # Version number as float
+        ('nOperationMode', 8, 'h'),         # Acquisition mode
         ('lActualAcqLength', 10, 'i'),
         ('nNumPointsIgnored', 14, 'h'),
-        ('lActualSweeps', 16, 'i'),
+        ('lActualSweeps', 16, 'i'),         # AKA lActualEpisodes
         ('lFileStartDate', 20, 'i'),
         ('lFileStartTime', 24, 'i'),
         ('lStopWatchTime', 28, 'i'),